+++ src/components/app/app.tsx
@@ -8,7 +8,7 @@ import { AppRoutes, AuthorizationStatus } from '../../const';
 import PrivateRoute from '../../components/private-route/private-route';
 import { FC } from 'react';
 import type { LocationItemProps } from '../../pages/main-screen/main-screen';
-import { Offer } from '../../types/offers';
+import { Offer } from '../../types/types';
 
 type AppProps = {
   offers: Offer[];
